Second série d'optimisation et de test. Correction d'un bug dans la discrétisation des éléments.

+++ trunk/src/lib/EF/EF_rigidite.h
@@ -24,8 +24,8 @@
 
 typedef struct
 {
-	int		noeudx;
-	int		noeudy;
+	long		noeudx;
+	long		noeudy;
 	double		matrice[6][6];
 } EF_rigidite;
